Create useToast hook. Moved toast logic out of plugin.
